Ensure action built
This adds a check in the CI to ensure that after running `yarn build` there are no unstaged changes in the repo.

It acts as a proof that the code in dist/index.js has not been tampered with and is indeed the bundled code of the action with its dependencies.