Use Markdown parser to detect thematic breaks
Using a regex matching for patterns similar to `---` is not enough because these could appear in code blocks where they shouldn't be considered as thematic breaks.

Tests related to the correct detection of all kinds of thematic breaks are not necessary since this is the responsibility of the Markdown parser we use.